fix sourcemaps on macos/temp dirs
When asking for a tempdir on macos, it returns `/var/folders`, and `/var` is actually a symlink to `/private/var`. This comes into play because somewhere along the way the `sourcemapFile` gets resolved to `/private/var/folders/*` while the input file is still `/var/folders/*`, and instead of the sourcemap referencing `./input.js`, it references `../../../../private/var/folders/.../input.js`. This update ensures that both input and `sourcemapFile` are resolved properly.